// expoFlix/src/types/api_movie.type.ts
//AQUI FICA A TIPAGEM DA NOSSA API.
//EXPORTE ESSA TIPAGEM
export type APIMOVIES = {
  adult: boolean;
  backdrop_path: string;
  genre_ids?: number[];
  genres?: Genres[];
  homepage?: string;
  id: number;
  imdb_id?: string;
  origin_country?: string[];
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies?: Production_companies[];
  production_countries?: Production_countries[];
  release_date: string;
  revenue?: number;
  runtime?: number;
  spoken_language?: Spoken_languages[];
  status?: string;
  tagline?: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};

export type Genres = {
  id: number;
  name: string;
};

export type Production_companies = {
  id: number;
  logo_path: string;
  name: string;
  origin_country: string;
};

export type Production_countries = {
  iso_3166_1: string;
  name: string;
};

export type Spoken_languages = {
  english_name: string;
  iso_639_1: string;
  name: string;
};